+++ src/components/Taxonomy/filter/SubjectFilters.jsx
@@ -8,9 +8,7 @@
 
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
-import {
-  StyledSubjectName,
-} from '../../style/LearningResourceTaxonomyStyles';
+import { StyledSubjectName } from '../../style/LearningResourceTaxonomyStyles';
 import FilterItem from './FilterItem';
 
 const SubjectFilters = ({
@@ -25,7 +23,7 @@ const SubjectFilters = ({
   const subject = structure.find(
     structureItem => structureItem.id === filterSubjectKey,
   );
-  
+
   if (!subject) {
     return null;
   }
