+++ src/containers/LearningResourcePage/components/LearningResourceTaxonomy.jsx
@@ -32,7 +32,7 @@ import {
   selectedResourceTypeValue,
 } from '../../../util/taxonomyHelpers';
 import handleError from '../../../util/handleError';
-import TopicConnections from './TopicConnections';
+import TopicConnections from './taxonomy/TopicConnections';
 import FilterConnections from '../../../components/Taxonomy/filter/FilterConnections';
 import SaveButton from '../../../components/SaveButton';
 import { FormActionButton } from '../../Form';
